pyproject.toml
[project]
name = "telegram-masumi-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx==0.25.2",
    "python-dotenv==1.0.0",
    "python-telegram-bot==20.7",
]
